// src/app/Model/allowancerelief.ts
export class AllowanceRelief {
      companyId: string; 
      divisionId: string; 
      departmentId: string; 
      fiscalYear:string 
      reliefTypeId: string;
      reliefRate:string 
      reliefAmount: string;
      lockedBy: string;
      lockTs: Date;
      branchCode: any; 
      data:string;
      status: string;
  
  }   